!Breaking: Use go-arg to parse generator arguments
Breaking change: the long flags in autometrics now take the long form
arguments starting with --

Also removes the unused `am-alertsgen` binary. Even though for the time
being we are using sloth to generate the alerts, at some point we might
want to change, and adding extra code/friction to do the change seems
unwarranted for now.