Add build information to the metrics/queries (#34)
* Add build information to the metrics/queries

This allows use cases where we can see and query the changes in metrics
after a version change in the code base.

Ref: https://github.com/autometrics-dev/autometrics-rs/pull/69
Ref: https://github.com/orgs/autometrics-dev/discussions/10

* Add --help and --version flags to the binary

Closes: #36

* !Breaking: Use go-arg to parse generator arguments

Breaking change: the long flags in autometrics now take the long form
arguments starting with --

Also removes the unused `am-alertsgen` binary. Even though for the time
being we are using sloth to generate the alerts, at some point we might
want to change, and adding extra code/friction to do the change seems
unwarranted for now.