Add support for ignoring/deleting generated links
The generator takes an extra argument, through command-line or
environment variable, that allows to opt-out of documentation
generation. There is also the possibility to opt-in this deactivation on
a per-function basis, using the `--no-doc` argument on the directive.

This also changes the name of the directive, from `autometrics:doc` to
`autometrics:inst`, as now the directive doesn't always add
documentation to the functions it instruments. For backwards
compatibility, the `doc` attribute is still accepted, and will be
removed in a later version.

Closes #38